+++ ls-bad-bot-blocker/CHANGES.md
@@ -1,3 +1,9 @@
+# 1.0-46
+* New domain page to enable/disable Bad Bot Blocker for that website
+* Generate per website block statistics
+* Generate block statistics for all websites (admin user only)
+* Added Bad Bot Blocker button in website Security section
+
 # 1.0-40
 Bug fixes
 + handle duplicate server_names_hash_max_size variable conflict if it exists